Update mock generator and xbox test app.
Updates the mock generator class to be more robust and allow other methods than get and websocket.

Updates the xbox test app to be the XboxWDPDriver.exe, which can store and retrieve credentials and default console, allowing this to feel like a true command line tool. Updates the XbUser.cmd script to show the functionality a bit. If both creds and console are stored now you can run "XbUser.cmd list" and it will print output very similar to "XbUser.exe list" :)

Also adds a /op:connect operation to the test app which will save the console as the default (and persist the WDP creds).