Updates the library for .NET to not accept untrusted roots.
There are two ways to override this--the connection can set a cert which could be provided by the app via the command line or the app can subscribe to an Untrusted cert event and have its own logic for determining if it trusts the certificate. The XboxWdpDriver app shows an example of both of these methods.

Currently the other .NET sample app just always accepts the cert. There will be a follow up change to fix this app to prompt the user to make a trust decision, as well as UWP needing a follow up change still to do the same thing.